clonotype.table {clonotypeR}	R Documentation

Create a table count of clonotypes or other features.

Description

Using a clonotype data frame loaded with read.clonotypes, clonotype.table will create a table counting how many times each clonotypes have been seen in each libraries. By default, the improductive rearrangements are filtered out.

Usage

clonotype.table(libs, feats=c("V","pep","J"), from=clonotypes, filter=from$improductive)

Arguments

libs
A character vector containing the name of one or many libraries. Same names must not appear twice.

feats
What to count. By default, it counts clonotypes, defined as c("V","pep","J"). But it can also count single features, such as the V or J segments.

from
Data frame as loaded by read.clonotypes. By default, clonotype.table will use a data frame called ‘clonotypes’.

filter
Logical vector to filter out clonotypes. By default it relies on the clonotypes data frame to provide a “improductive” column, that indicates clonotypes with a stop codon or a frame shift.

Value

clonotype.table returns a data frame, where row names are features (clonotypes, segment names, ...), column names are libraries, and values are number of times each feature was found in each library.

Author(s)

Charles Plessy

See Also

read.clonotypes

Examples

# Read the package's example data
clonotypes <- read.clonotypes(system.file("data/clonotypes.txt.gz", package = "clonotypeR"))

# Count J segments
j <- clonotype.table(levels(clonotypes$lib), "J")

# Normalise counts in parts per million
J <- data.frame(prop.table(as.matrix(j),2) * 1000000)

[Package clonotypeR version 0.1 Index]