extdata {clonotypeR}	R Documentation

Extra data used to calculate ID numbers in Yassai et al.'s nomenclature.

Description

Data frame derived from Table 1 of Yassai et al., 2009, to construct clonotype names.

Details

V_after_C: sequence of the V segments after their conserved cystein.

J_before_FGxG: sequence of the J segments before their conserved FGxG motif.

codon_ids: data frame derived from Table 1 of Yassai et al., 2009, to construct clonotype names.

The V_after_C and J_before_FGxG tables are, generated from the mouse reference data with the command: make refresh-data in the source repository of clonotypeR.

Value

codon_ids:

codon
Nucleotide triplet.

aminoacid
Single-letter amino acid abbreviation (“O” for stop).

id
ID numbers assigned to the codons for each amino acids.

J_before_FGxG:

row name
J segment name, for instance “TRAJ61”.

sequence
Sequence of the nucleic acids preceding the first codon of the conserved FGxG motif.

V_after_C:

row name
V segment name, for instance “TRAV1”.

sequence
Sequence of the nucleic acids following the codon of the conserved cysteine.

References

A clonotype nomenclature for T cell receptors. Maryam B. Yassai, Yuri N. Naumov, Elena N. Naumova and Jack Gorski Immunogenetics, 2009, Volume 61, Number 7, Pages 493-502

See Also

yassai_identifier

Examples

V_after_C <- read.table(system.file('extdata', 'V_after_C.txt.gz', package = "clonotypeR"), stringsAsFactors=FALSE)

J_before_FGxG <- read.table(system.file('extdata', 'J_before_FGxG.txt.gz', package = "clonotypeR"), stringsAsFactors=FALSE)

codon_ids <- read.table(system.file('extdata', 'codon_ids.txt.gz', package = "clonotypeR"), header=TRUE, row.names=1)

[Package clonotypeR version 1.1.3 Index]