common_clonotypes {clonotypeR}	R Documentation

Reports clonotypes common between samples.

Description

When given two groups of samples, lists the clonotypes that have been observed at least in one library of each group. The returned list can be used to subset a data frame produced by clonotype.table.

When given a table of clonotypes, produces a matrix in which each cell reports the number of clonotypes found in a pair of libraries.

Usage

common_clonotypes(group1, group2, mode="groups", data=clonotypes)

Arguments

group1
A character vector containing clonotype library names

group2
A character vector containing clonotype library names

mode
By default, “groups” when “group1” or “group2” is provided, “matrix” mode otherwise.

data
A clonotype table where the data is stored. Default: “clonotypes”.

Value

In “groups” mode, a character vector of clontoype names. Their order follows the original row name order of the clonotype table.

In “matric” mode, a symmetric matrix of numbers, with the same row and column names as the row names of the clonotype table used.

Author(s)

Charles Plessy

See Also

clonotype.table, unique_clonotypes

Examples

# Load example data
clonotypes.long <- read.clonotypes(system.file("data/clonotypes.txt.gz", package = "clonotypeR"))
clonotypes <- clonotype.table(levels(clonotypes.long$lib), from=clonotypes.long)
summary(clonotypes)

# List clonotypes found in library A, and B or C.
common_clonotypes(group1="A", group2=c("B","C"))

# Count clonotypes found in library A, and B or C.
length(common_clonotypes(group1="A", group2=c("B","C")))

# Matrix of numbers of common clonotypes
common_clonotypes()

[Package clonotypeR version 0.1 Index]