J_before_FGxG {clonotypeR}	R Documentation

Sequence of the J segments before their conserved FGxG motif.

Description

Table intended for internal use in the yassai.nomenclature function, generated from the reference data with the command: make refresh-data.

Usage

data(J_before_FGxG)

Value

row name
J segment name, for instance “TRAJ61”.

sequence
Sequence of the nucleic acids preceding the first codon of the conserved FGxG motif.

See Also

yassai.nomenclature

[Package clonotypeR version 0.9 Index]