V_after_C {clonotypeR}	R Documentation

Sequence of the V segments after their conserved cystein.

Description

Table intended for internal use in the yassai_identifier function, generated from the reference data with the command: make refresh-data.

Usage

data(V_after_C)

Value

row name
V segment name, for instance “TRAV1”.

sequence
Sequence of the nucleic acids following the codon of the conserved cysteine.

See Also

yassai_identifier

[Package clonotypeR version 0.99.1 Index]