is_unproductive {clonotypeR}	R Documentation

Determines if clonotype sequences are productive.

Description

ClonotypeR identifies V and J segments, isolates the DNA sequence between the conserved cystein and the FGxG motifs, and translates it. This functions verifies that this sequence is in frame and has no stop codon.

Usage

is_unproductive(data)

Arguments

data
Data frame of clonotype sequences, or character vector describing a single clonotype, where the DNA sequence is available under the name “dna” and its translation available under the name “pep”.

Details

Clonotypes are marked unproductive if the length of their DNA sequence is not a multiple of 3, or if they contain a stop codon, as indicated by an asterisk in the translated sequence.

Value

Returns a logical vector, with one value per row in the original data.

Author(s)

Charles Plessy

See Also

read_clonotypes

Examples

clonotypes <- read_clonotypes(system.file('extdata', 'clonotypes.txt.gz', package = "clonotypeR"))
is_unproductive(clonotypes)

[Package clonotypeR version 1.1.3 Index]