read_clonotypes {clonotypeR}	R Documentation

Reads a clonotype_table and returns a data frame.

Description

Reads a clonotype_table in a TAB-separated or OSCT format, and returns a data frame that has eight columns, for library name, V and J segments names, sequence read identifier, DNA, sequence quality, aminoacid sequence of the CDR3 region, mark for unproductive recombinations, and mark for ambiguous sequences.

Usage

read_clonotypes(filename, scores=TRUE, ...)

Arguments

filename
Path to the tabulation-delimited text file containing the extracted clonotypes.

scores
Set to false to load legacy data that did not contain “score” and “mapq” columns.

...
The rest of the arguments are passed to the read.table() function.

Value

lib
Library name (factor).

V
V segment name (factor).

J
J segment name (factor).

score
Alignment score (numeric).

mapq
Mapping quality (numeric). A sequence with a good alignment score will still have a low mapping quality if there are good alternative alignments to other V segments.

read
Sequence read identifier (character).

dna
DNA sequence of the CDR3 region (character).

qual
Quality values for the DNA sequence (character).

pep
Translation of the DNA sequence (character).

unproductive
Flag indicating stop codons or frame shifts (logical).

ambiguous
Flag indicating that the DNA sequences has ambiguous (“N”) nucleotides (logical).

Author(s)

Charles Plessy

See Also

clonotype_table, is_unproductive, read.table, Order Switchable Column Table (OSCT, http://sourceforge.net/projects/osctf/)

Examples

clonotypes <- read_clonotypes(system.file('extdata', 'clonotypes.txt.gz', package = "clonotypeR"))

[Package clonotypeR version 1.1.3 Index]