unique_clonotypes {clonotypeR}	R Documentation

lists unique clonotypes in libraries.

Description

Finds all the clonotypes expressed in one or more libraries, and returns a vector where they are listed once. This vector can be used to subset a clonotype table.

Usage

unique_clonotypes(libs, from=clonotypes)

Arguments

libs
A character vector contain clonotype library names

from
The name of the clonotype table where the data is stored.

Details

Uses the table called “clonotypes” by default.

Value

Character vector of clontoype names. Their order follows the original row name order of the clonotype table.

Author(s)

Charles Plessy

See Also

clonotype.table, common_clonotypes

Examples

# Load example data
clonotypes.long <- read.clonotypes(system.file("data/clonotypes.txt.gz", package = "clonotypeR"))
clonotypes <- clonotype.table(levels(clonotypes.long$lib), from=clonotypes.long)
summary(clonotypes)

# List clonotypes found in library A.
unique_clonotypes("A")

# List clonotypes found in library A or B.
unique_clonotypes(c("A","B"))

[Package clonotypeR version 0.1 Index]