yassai_identifier {clonotypeR}	R Documentation

TCR clonotype identifier (Yassai et al.).

Description

The clonotype nomenclature defined by Yassai et al. in http://dx.doi.org/10.1007/s00251-009-0383-x.

Usage

yassai_identifier(data, V_after_C, J_before_FGxG)

Arguments

data
A data frame or a character vector containing a clonotype(s) with proper row or element names.

V_after_C
(optional) A data frame indicating the aminoacids following the conserved cystein for each V segment.

J_before_FGxG
(optional) A data frame indicating the aminoacids preceding the conserved FGxG motif for each V segment.

Details

By default, codeyassai_identifier() assume mouse sequences and will load the V_after_C and J_before_FGxG tables distributed in this package. It is possible to provide alternative tables either by passing them directly as argument, or by installing them as “./inst/extdata/V_after_C.txt.gz” and “./inst/extdata/J_before_FGxG.txt.gz”.

Value

The name (for instance sIRSSy.1456B19S1B27L11) consists of five segments:

CDR3 amino acid identifier (ex. sIRSSy), followed by a dot ;

CDR3 nucleotide sequence identifier (ex. 1456) ;

variable (V) segment identifier (ex. BV19S1) ;

joining (J) segment identifier (ex. BJ2S7) ;

CDR3 length identifier (ex. L11).

Author(s)

Charles Plessy

See Also

codon_ids, J_before_FGxG, V_after_C

Examples

clonotypes <- read_clonotypes(system.file('extdata', 'clonotypes.txt.gz', package = "clonotypeR"))
head(yassai_identifier(clonotypes))

[Package clonotypeR version 0.99.4 Index]