unique_clonotypes {clonotypeR}	R Documentation

Lists unique clonotypes in libraries.

Description

Finds all the clonotypes expressed in one or more libraries, and returns a vector where they are listed once. This vector can be used to subset a clonotype_table.

Usage

unique_clonotypes(..., data)

Arguments

...
One or more character vectors contain clonotype library names

data
The name of the clonotype_table where the data is stored.

Details

Uses the table called “clonotypes” by default.

Value

Character vector of clonotype names. Their order follows the original row name order of the clonotype_table.

Author(s)

Charles Plessy

See Also

clonotype_table, common_clonotypes

Examples

# Load example data
clonotypes.long <- read_clonotypes(system.file('extdata', 'clonotypes.txt.gz', package = "clonotypeR"))
clonotypes <- clonotype_table(levels(clonotypes.long$lib), data=clonotypes.long)
summary(clonotypes)

# List clonotypes found in library A.
unique_clonotypes("A", data=clonotypes)

# List clonotypes found in library A or B.
unique_clonotypes("A","B", data=clonotypes)

[Package clonotypeR version 1.1.3 Index]